requirements.txt
runpod>=1.7.0
torch>=2.4.0
git+https://github.com/huggingface/diffusers.git
transformers>=4.45.0
accelerate>=0.34.0
qwen-vl-utils
Pillow>=10.0.0
